Fix of a possible segfault if fstab file is empty in do_fstrim_filesystems
This patch fixes an segfault in case of an empty fstab file in the filesystem.
The corresponding fstab struct in memory is NULL in that case and
therefore the we now chack that before do_fstrim_filesystems() method is called
and starting to iterate over the fstab structure.

This fixes:
Jira-> TM-274 (lollipop 5.1.0/5.1.1 fstrim crash during startup of a0 and a1)

Change-Id: I933ae5d85412ab3c97b3e5d34ac4b89a41a9130d